--- apps/control/src/utils/scheduling/MaterialTableUtils.ts
@@ -1,819 +1,0 @@
-// Columns 정의
-import { ColumnDataType } from '@postcoil/ui/config/TableConfig';
-
-import { MaterialDataType } from '@/config/scheduling/ContentConfig';
-import { MaterialDTO } from '@/config/scheduling/DTO';
-// import { Tag } from 'antd';
-
-export const mockcolumns: ColumnDataType<MaterialDataType>[] = [
-  {
-    title: '재료번호',
-    dataIndex: 'materialNo',
-    key: 'materialNo',
-    // fixed: true,
-  },
-  { title: '공정코드', dataIndex: 'currProc', key: 'currProc' },
-  {
-    title: '롤단위',
-    dataIndex: 'rollUnit',
-    key: 'rollUnit',
-    sortable: true,
-    otherProps: {
-      filters: [
-        {
-          text: 'A',
-          value: 'A',
-        },
-        {
-          text: 'B',
-          value: 'B',
-        },
-      ],
-      onFilter: (value: string, record: { rollUnit: string[] }) =>
-        record.rollUnit.indexOf(value) === 0,
-    },
-  },
-  { title: '폭', dataIndex: 'width', key: 'width', sortable: true },
-  {
-    title: '두께',
-    dataIndex: 'thickness',
-    key: 'thickness',
-    sortable: true,
-  },
-  {
-    title: '온도',
-    dataIndex: 'temperature',
-    key: 'temperature',
-    sortable: true,
-  },
-  {
-    title: '차공정',
-    dataIndex: 'nextProc',
-    key: 'nextProc',
-    sortable: true,
-    otherProps: {
-      filters: [
-        {
-          text: 'CGL',
-          value: '1CGL' || '2CGL',
-        },
-        {
-          text: 'EGL',
-          value: '1EGL' || '2EGL',
-        },
-        {
-          text: '포장',
-          value: '101' || '201',
-        },
-      ],
-      onFilter: (value: string, record: { nextProc: string[] }) =>
-        record.nextProc.indexOf(value) === 0,
-    },
-  },
-  {
-    title: '목표폭',
-    dataIndex: 'goalWidth',
-    key: 'goalWidth',
-    sortable: true,
-    defaultSortOrder: 'descend',
-  },
-  {
-    title: '목표두께',
-    dataIndex: 'goalThickness',
-    key: 'goalThickness',
-    sortable: true,
-  },
-];
-
-export const transformedData = (data: MaterialDTO[]): MaterialDataType[] => {
-  return data.map((item) => ({
-    key: item.id,
-    materialNo: item.materialNo,
-    status: item.status,
-    // opCode: item.opCode,
-    currProc: item.currProc,
-    temperature: item.temperature,
-    // type: item.type,
-    // progress: item.progress,
-    rollUnit: item.rollUnit,
-    // outerDia: item.outerDia,
-    // innerDia: item.innerDia,
-    width: item.width,
-    // length: item.length,
-    thickness: item.thickness,
-    // weight: item.weight,
-    // passProc: item.passProc,
-    // preProc: item.preProc,
-    nextProc: item.nextProc,
-    // storageLoc: item.storageLoc,
-    // yard: item.yard,
-    goalWidth: item.goalWidth,
-    goalThickness: item.goalThickness,
-  }));
-};
-
-// Table Data 정의
-export const mockdata: MaterialDataType[] = [
-  {
-    key: '1',
-    materialNo: 'C6EK0001',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 35.9824301584831,
-    innerDia: 22,
-    width: 775,
-    length: 2547,
-    thickness: 0.621409153683,
-    weight: 68.5737101624795,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    storageLoc: '7-2-20-2-1',
-    yard: '1A1B',
-    goalWidth: 770,
-    goalThickness: 0.48,
-  },
-  {
-    key: '2',
-    materialNo: 'C3DL0002',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 52.9602313116795,
-    innerDia: 25,
-    width: 772,
-    length: 6848,
-    thickness: 0.504473434969,
-    weight: 434.688234352393,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    changed: true,
-    goalWidth: 772,
-    goalThickness: 0.49,
-  },
-  {
-    key: '3',
-    materialNo: 'C3i00003',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 49.4609835245515,
-    innerDia: 24,
-    width: 772,
-    length: 5876,
-    thickness: 0.620238096071,
-    weight: 645.773619063763,
-    passProc: '1CAL',
-    preProc: '',
-    nextProc: '포장',
-    goalWidth: 770,
-    goalThickness: 0.57,
-  },
-  {
-    key: '4',
-    materialNo: 'C3Nj0004',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 54.0121547172835,
-    innerDia: 28,
-    width: 769,
-    length: 6702,
-    thickness: 0.337581205233,
-    weight: 666.443669362022,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    goalWidth: 765,
-    goalThickness: 0.3,
-  },
-  {
-    key: '5',
-    materialNo: 'Cbq0005',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 60.8096102448159,
-    innerDia: 28,
-    width: 766,
-    length: 9154,
-    thickness: 0.50512372108997,
-    weight: 504.379666723385,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '1EGL',
-    goalWidth: 765,
-    goalThickness: 0.39,
-  },
-  {
-    key: '6',
-    materialNo: 'Cow0006',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'B',
-    outerDia: 53.7773466191942,
-    innerDia: 25,
-    width: 768,
-    length: 7122,
-    thickness: 0.6426305159542,
-    weight: 217.257526025889,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    goalWidth: 765,
-    goalThickness: 0.62,
-  },
-  {
-    key: '7',
-    materialNo: 'COz0007',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 42.1349622390059,
-    innerDia: 21,
-    width: 774,
-    length: 4192,
-    thickness: 0.51703509387288,
-    weight: 122.689476558989,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    goalWidth: 772,
-    goalThickness: 0.47,
-  },
-  {
-    key: '8',
-    materialNo: 'C3Lp0008',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 53.3285872359957,
-    innerDia: 25,
-    width: 775,
-    length: 6971,
-    thickness: 0.714454085291,
-    weight: 417.961959629394,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    goalWidth: 772,
-    goalThickness: 0.58,
-  },
-  {
-    key: '9',
-    materialNo: 'CGW0009',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'B',
-    outerDia: 53.526933092068,
-    innerDia: 29,
-    width: 784,
-    length: 6359,
-    thickness: 0.70274970271302,
-    weight: 105.774351816828,
-    passProc: '1PCM',
-    preProc: '1CAL',
-    nextProc: '1EGL',
-    goalWidth: 765,
-    goalThickness: 0.67,
-  },
-  {
-    key: '10',
-    materialNo: 'Piy0010',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'B',
-    outerDia: 40.344223871778,
-    innerDia: 30,
-    width: 737,
-    length: 2286,
-    thickness: 0.86932033936218,
-    weight: 33.9806455809316,
-    passProc: '1EGL',
-    preProc: '1CAL',
-    nextProc: '1CGL',
-    goalWidth: 735,
-    goalThickness: 0.7,
-  },
-];
-
-export const schMockcolumns: ColumnDataType<MaterialDataType>[] = [
-  {
-    title: '재료번호',
-    dataIndex: 'materialNo',
-    key: 'materialNo',
-  },
-  // { title: '스케줄명', dataIndex: 'scheduleName', key: 'scheduleName' },
-  { title: '공정코드', dataIndex: 'curProc', key: 'curProc' },
-  {
-    title: '재료단위',
-    dataIndex: 'rollUnit',
-    key: 'rollUnit',
-    sortable: true,
-  },
-  { title: '폭', dataIndex: 'width', key: 'width', sortable: true },
-  {
-    title: '길이',
-    dataIndex: 'length',
-    key: 'length',
-    sortable: true,
-  },
-  {
-    title: '두께',
-    dataIndex: 'thickness',
-    key: 'thickness',
-    sortable: true,
-  },
-  {
-    title: '차공정',
-    dataIndex: 'nextProc',
-    key: 'nextProc',
-    sortable: true,
-    otherProps: {
-      filters: [
-        {
-          text: '1CGL',
-          value: '1CGL',
-        },
-        {
-          text: '1EGL',
-          value: '1EGL',
-        },
-        {
-          text: '포장',
-          value: '포장',
-        },
-      ],
-      onFilter: (value: string, record: { nextProc: string[] }) =>
-        record.nextProc.indexOf(value) === 0,
-    },
-  },
-  {
-    title: '목표폭',
-    dataIndex: 'goalWidth',
-    key: 'goalWidth',
-    sortable: true,
-    // defaultSortOrder: 'descend',
-  },
-  {
-    title: '목표두께',
-    dataIndex: 'goalThickness',
-    key: 'goalThickness',
-    sortable: true,
-  },
-  {
-    title: '작업소요시간',
-    dataIndex: 'worktime',
-    key: 'worktime',
-    sortable: true,
-  },
-  {
-    title: 'Reject여부',
-    dataIndex: 'reject',
-    key: 'reject',
-    // TODO: tag로 reject 여부 디자인!
-    // render: (_, record) => {
-    //   let color = 'green';
-    //   let tag = '';
-    //   console.log(record);
-    //   if (record.reject === 'T') {
-    //     color = 'volcano';
-    //     tag = 'reject';
-    //   }
-    //   return (
-    //     <Tag color={color} key={color}>
-    //       {tag.toUpperCase()}
-    //     </Tag>
-    //   );
-    // },
-  },
-];
-
-// Table Data 정의
-export const schMockdataA: MaterialDataType[] = [
-  {
-    key: '1',
-    materialNo: 'C3Lp0008',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 53.3285872359957,
-    innerDia: 25,
-    width: 775,
-    length: 6971,
-    thickness: 0.714454085291,
-    weight: 417.961959629394,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    goalWidth: 772,
-    goalThickness: 0.58,
-    worktime: 35,
-  },
-  {
-    key: '2',
-    materialNo: 'COz0007',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 42.1349622390059,
-    innerDia: 21,
-    width: 774,
-    length: 4192,
-    thickness: 0.51703509387288,
-    weight: 122.689476558989,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    goalWidth: 772,
-    goalThickness: 0.47,
-    changed: true,
-    worktime: 48,
-    reject: 'T',
-  },
-  {
-    key: '3',
-    materialNo: 'C3DL0002',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 52.9602313116795,
-    innerDia: 25,
-    width: 775,
-    length: 6848,
-    thickness: 0.504473434969,
-    weight: 434.688234352393,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    goalWidth: 772,
-    goalThickness: 0.49,
-    worktime: 29,
-  },
-
-  {
-    key: '4',
-    materialNo: 'C6EK0001',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 35.9824301584831,
-    innerDia: 22,
-    width: 775,
-    length: 2547,
-    thickness: 0.621409153683,
-    weight: 68.5737101624795,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    storageLoc: '7-2-20-2-1',
-    yard: '1A1B',
-    orderNumber: '20240730-F6BG',
-    goalWidth: 770,
-    goalThickness: 0.48,
-    worktime: 34,
-  },
-  {
-    key: '5',
-    materialNo: 'C3i00003',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 49.4609835245515,
-    innerDia: 24,
-    width: 772,
-    length: 5876,
-    thickness: 0.620238096071,
-    weight: 645.773619063763,
-    passProc: '1CAL',
-    preProc: '',
-    nextProc: '포장',
-    goalWidth: 770,
-    goalThickness: 0.57,
-    worktime: 38,
-  },
-
-  {
-    key: '6',
-    materialNo: 'Cbq0005',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 60.8096102448159,
-    innerDia: 28,
-    width: 766,
-    length: 9154,
-    thickness: 0.50512372108997,
-    weight: 504.379666723385,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '1EGL',
-    goalWidth: 765,
-    goalThickness: 0.39,
-    worktime: 35,
-  },
-  {
-    key: '7',
-    materialNo: 'C3Nj0004',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 54.0121547172835,
-    innerDia: 28,
-    width: 769,
-    length: 6702,
-    thickness: 0.337581205233,
-    weight: 666.443669362022,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    goalWidth: 765,
-    goalThickness: 0.3,
-    worktime: 30,
-  },
-];
-
-// Table Data 정의
-export const schMockdataA2: MaterialDataType[] = [
-  {
-    key: '1',
-    materialNo: 'C3Lp0008',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 53.3285872359957,
-    innerDia: 25,
-    width: 775,
-    length: 6971,
-    thickness: 0.714454085291,
-    weight: 417.961959629394,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    goalWidth: 772,
-    goalThickness: 0.58,
-    worktime: 35,
-  },
-  {
-    key: '2',
-    materialNo: 'COz0007',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 42.1349622390059,
-    innerDia: 21,
-    width: 774,
-    length: 4192,
-    thickness: 0.51703509387288,
-    weight: 122.689476558989,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    goalWidth: 772,
-    goalThickness: 0.47,
-    worktime: 48,
-  },
-  {
-    key: '3',
-    materialNo: 'C3DL0002',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 52.9602313116795,
-    innerDia: 25,
-    width: 775,
-    length: 6848,
-    thickness: 0.504473434969,
-    weight: 434.688234352393,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    changed: true,
-    goalWidth: 772,
-    goalThickness: 0.49,
-    worktime: 29,
-  },
-
-  {
-    key: '4',
-    materialNo: 'C6EK0001',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 35.9824301584831,
-    innerDia: 22,
-    width: 775,
-    length: 2547,
-    thickness: 0.621409153683,
-    weight: 68.5737101624795,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    storageLoc: '7-2-20-2-1',
-    yard: '1A1B',
-    orderNumber: '20240730-F6BG',
-    goalWidth: 770,
-    goalThickness: 0.48,
-    worktime: 34,
-  },
-  {
-    key: '5',
-    materialNo: 'C3i00003',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 49.4609835245515,
-    innerDia: 24,
-    width: 772,
-    length: 5876,
-    thickness: 0.620238096071,
-    weight: 645.773619063763,
-    passProc: '1CAL',
-    preProc: '',
-    nextProc: '포장',
-    goalWidth: 770,
-    goalThickness: 0.57,
-    worktime: 38,
-  },
-
-  {
-    key: '6',
-    materialNo: 'Cbq0005',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 60.8096102448159,
-    innerDia: 28,
-    width: 766,
-    length: 9154,
-    thickness: 0.50512372108997,
-    weight: 504.379666723385,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '1EGL',
-    goalWidth: 765,
-    goalThickness: 0.39,
-    worktime: 35,
-  },
-  {
-    key: '7',
-    materialNo: 'C3Nj0004',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'A',
-    outerDia: 54.0121547172835,
-    innerDia: 28,
-    width: 769,
-    length: 6702,
-    thickness: 0.337581205233,
-    weight: 666.443669362022,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    goalWidth: 765,
-    goalThickness: 0.3,
-    worktime: 30,
-  },
-];
-
-// Table Data 정의
-export const schMockdataB: MaterialDataType[] = [
-  {
-    key: '6',
-    materialNo: 'Cow0006',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'B',
-    outerDia: 53.7773466191942,
-    innerDia: 25,
-    width: 768,
-    length: 7122,
-    thickness: 0.6426305159542,
-    weight: 217.257526025889,
-    passProc: '1CAL',
-    preProc: '1PCM',
-    nextProc: '포장',
-    goalWidth: 765,
-    goalThickness: 0.62,
-  },
-
-  {
-    key: '9',
-    materialNo: 'CGW0009',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'B',
-    outerDia: 53.526933092068,
-    innerDia: 29,
-    width: 784,
-    length: 6359,
-    thickness: 0.70274970271302,
-    weight: 105.774351816828,
-    passProc: '1PCM',
-    preProc: '1CAL',
-    nextProc: '1EGL',
-    goalWidth: 765,
-    goalThickness: 0.67,
-  },
-  {
-    key: '10',
-    materialNo: 'Piy0010',
-    status: '2',
-    opCode: 'C',
-    curProc: '1CAL',
-    type: 'C',
-    progress: 'D',
-    rollUnit: 'B',
-    outerDia: 40.344223871778,
-    innerDia: 30,
-    width: 737,
-    length: 2286,
-    thickness: 0.86932033936218,
-    weight: 33.9806455809316,
-    passProc: '1EGL',
-    preProc: '1CAL',
-    nextProc: '1CGL',
-    goalWidth: 735,
-    goalThickness: 0.7,
-  },
-];
